backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["takuya.mouri <ayukat101699@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
uvicorn = "^0.34.2"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
